+++ src/main/java/roomescape/dto/reservationtime/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.reservationtime;
 
 import java.time.LocalTime;
